+++ src/modules/demux/module_demux.cpp
@@ -20,11 +20,28 @@ void module_demux::run( options *opts )
 
     std::string index_str;
 
-    const std::size_t forward_start  = std::get<0>( d_opts->index1_data );
-    const std::size_t forward_length = std::get<1>( d_opts->index1_data );
+    // fif use case
+    std::vector<flex_idx> flexible_idx_data;
+    std::vector<sequential_map<sequence, sample>::iterator> idx_match_list;
 
-    const std::size_t reverse_start  = std::get<0>( d_opts->index2_data );
-    const std::size_t reverse_length = std::get<1>( d_opts->index2_data );
+    if( d_opts->flexible_idx_fname.empty() )
+        {
+            flexible_idx_data.emplace_back( d_opts->sample_indexes[0],            // index name
+                                            "r1",                                 // read direction
+                                            std::get<0>( d_opts->index1_data ),   // start
+                                            std::get<1>( d_opts->index1_data ),   // length
+                                            std::get<2>( d_opts->index1_data ) ); // num mismatch
+            flexible_idx_data.emplace_back( d_opts->sample_indexes[1],
+                                            "r2",
+                                            std::get<0>( d_opts->index2_data ),
+                                            std::get<1>( d_opts->index2_data ),
+                                            std::get<2>( d_opts->index2_data ));
+        }
+    else
+        {
+            fif_parser flex_parser;
+            flexible_idx_data = flex_parser.parse( d_opts->flexible_idx_fname );
+        }
 
 
     std::size_t read_index = 0;
@@ -33,8 +50,8 @@ void module_demux::run( options *opts )
     parallel_map<sequence, std::size_t> non_perfect_match_seqs;
 
     sequential_map<sequence, sample> index_map;
-    std::map<std::pair<std::string,std::string>,
-             std::pair<std::string,std::vector<std::size_t>>> diagnostic_map;
+    using barcode = std::pair<std::string, std::vector<std::size_t>>;
+    std::map<std::string,std::vector<barcode>> diagnostic_map;
     omp_set_num_threads( opts->num_threads );
 
     total_time.start();;
@@ -52,9 +69,15 @@ void module_demux::run( options *opts )
     // parse samplelist
     samplelist_parser samplelist_p;
 
+    if( !d_opts->flexible_idx_fname.empty() )
+        {
+            for( std::size_t curr_row = 0; curr_row < flexible_idx_data.size(); ++curr_row )
+                {
+                    d_opts->sample_indexes.emplace_back( flexible_idx_data[curr_row].idx_name );
+                }
+        }
     std::vector<sample> samplelist = samplelist_p.parse( d_opts );
 
-    // open the read file so we can iterate over it
     std::ifstream reads_file( d_opts->input_r1_fname, std::ios_base::in );
     std::ifstream r2_reads;
 
@@ -80,7 +103,7 @@ void module_demux::run( options *opts )
 
     index_seqs     = fasta_p.parse( d_opts->index_fname );
 
-    create_index_map( index_map, index_seqs, samplelist );
+    create_index_map( index_map, index_seqs, samplelist, flexible_idx_data );
     sequential_map<sequence, sample> sample_table;
     sequential_map<sequence,sample>::size_type num_samples = index_map.size();
 
@@ -101,15 +124,10 @@ void module_demux::run( options *opts )
     sequence_indexer lib_idx;
     sequence_indexer index_idx;
 
-    // index the forward index sequences and the library
-    // sequences
     lib_idx.index( library_seqs );
     index_idx.index( index_seqs );
 
     std::size_t sample_id = 0;
-    sequential_map<sequence, sample>::iterator idx2_match;
-    sequential_map<sequence, sample>::iterator idx1_match;
-
     std::istream *reads_ptr = &reads_file;
     std::istream *r2_reads_ptr = &r2_reads;
 
@@ -151,87 +169,72 @@ void module_demux::run( options *opts )
     #endif
         }
 
-
     std::istream& reads_file_ref = *reads_ptr;
     std::istream& r2_reads_ref   = *r2_reads_ptr;
-    std::size_t idx1_match_total = 0;
-    std::size_t idx2_match_total = 0;
+    std::vector<std::pair<std::string,size_t>> index_match_totals;
     std::size_t var_reg_match_total = 0;
+
+    for( const auto& curr_index : flexible_idx_data )
+        {
+            index_match_totals.emplace_back( std::make_pair( curr_index.idx_name, 0 ) );
+        }
 
     if( !d_opts->diagnostic_fname.empty() )
         {
-            for( const auto index : index_map )
+            for( const auto sample : samplelist )
                 {
-                    diagnostic_map.insert( { { index.second.string_ids.first, index.second.string_ids.second },
-                                             { index.second.name , std::vector<std::size_t>{0,0,0} } } );
+                    std::vector<barcode> barcodes;
+                    for( std::size_t curr_barcode = 0; curr_barcode < sample.string_ids.size(); ++curr_barcode )
+                        {
+                            barcodes[curr_barcode] = std::make_pair( sample.string_ids[curr_barcode], std::vector<std::size_t>( sample.string_ids.size(), 0 ) );
+                        }
+                    diagnostic_map.emplace( sample.name, barcodes );
                 }
         }
-    // while we still have reads to process, loop terminates when no fastq records
-    // are read from the file
+
     while( fastq_p.parse( reads_file_ref, reads, d_opts->read_per_loop  ) )
         {
 
             if( r2_reads_included )
                 {
-                    // get the index sequences for this set of sequences
                     fastq_p.parse( r2_reads_ref, r2_seqs, d_opts->read_per_loop );
                 }
 
            #pragma omp parallel for private( seq_iter, nuc_seq, read_index, index_str, adapter, sample_id,  \
                                               idx2_match, idx1_match ) \
                shared( seq_start, seq_length, d_opts, num_samples, reference_counts, library_seqs, index_seqs, r2_seqs ) \
-                reduction( +:processed_total, idx1_match_total, idx2_match_total, var_reg_match_total, processed_success, concatemer_found ) \
+                reduction( +:processed_total, index_match_totals, var_reg_match_total, processed_success, concatemer_found ) \
                 schedule( dynamic )
 
             for( read_index = 0; read_index < reads.size(); ++read_index )
                 {
-                    // lambda to return whether a match was found
-                    // if no reverse indexes are used, then we only
-                    // check the forward index match, otherwise
-                    // we need to check both
+                    std::size_t curr_index = 0;
                     auto match_found = [&]() -> bool
                         {
-                            if( idx1_match != index_map.end() )
+                            bool found_match = true;
+                            std::string concat_idx_seq = "";
+                            while( found_match && curr_index < idx_match_list.size() )
                                 {
-                                    ++idx1_match_total;
-                                    // When diagnostic file output is included, increment f & r index for associated sample name
-                                    if( !d_opts->diagnostic_fname.empty() )
+                                    // determine which indexes found matches
+                                    auto index_match = idx_match_list[curr_index];
+                                    if( index_match != index_map.end() )
                                         {
-                                            for( auto& sample : diagnostic_map )
-                                                {
-                                                    if( idx1_match->first.name == sample.first.first )
-                                                        {
-                                                            sample.second.second.at( 0 ) += 1;
-                                                        }
-                                                }
+                                            concat_idx_seq.append( index_match->first.seq );
+                                            index_match_totals[curr_index].second++;
                                         }
-
-                                    if( reverse_length == 0 )
+                                    else
                                         {
-                                            return true;
+                                            found_match = false;
                                         }
-                                    if( reverse_length != 0
-                                        && idx2_match != index_map.end() )
-                                        {
-                                            std::string index_seq_concat = idx1_match->first.seq + idx2_match->first.seq;
-                                            if( index_map.find( sequence("", index_seq_concat ) ) != index_map.end() )
-                                                ++idx2_match_total;
-                                            for( auto& sample : diagnostic_map )
-                                                {
-                                                    if( idx2_match->first.name == sample.first.second
-                                                        && idx1_match->first.name == sample.first.first )
-                                                        {
-                                                            sample.second.second.at( 1 ) += 1;                                                            
-                                                        }
-                                                }
-                                            return true;
-                                        }
+                                    ++curr_index;
                                 }
-
+                            if( index_map.find( sequence( "", concat_idx_seq ) ) != index_map.end() )
+                                {
+                                    return true;
+                                }
                             return false;
                         };
 
-                    // checks to see that a concatemer was included
                     auto found_concatemer = [&]() -> bool
                         {
                             return
@@ -255,31 +258,37 @@ void module_demux::run( options *opts )
                                         )
                                    );
                         };
-
-                    if( reverse_length > 0 )
+                    for( const auto& index : flexible_idx_data )
                         {
-                            if( r2_seqs.size() == 0 )
+                            sequential_map<sequence, sample>::iterator index_match;
+                            if( index.read_name == "r1" )
                                 {
-                                    idx2_match = _find_with_shifted_mismatch( index_map, reads[ read_index ],
-                                                                               index_idx, std::get<2>( d_opts->index2_data ),
-                                                                               reverse_start, reverse_length
-                                                                             );
-                            }
+                                    index_match = _find_with_shifted_mismatch( index_map, reads[ read_index ],
+                                                                            index_idx, index.num_mismatch,
+                                                                            index.idx_start, index.idx_len
+                                                                            );
+                                }
                             else
                                 {
-                                    idx2_match = _find_with_shifted_mismatch( index_map, r2_seqs[ read_index ],
-                                                                               index_idx, std::get<2>( d_opts->index2_data ),
-                                                                               reverse_start, reverse_length
-                                                                             );
+                                    if( r2_seqs.size() == 0 )
+                                        {
+                                            index_match = _find_with_shifted_mismatch( index_map, reads[ read_index ],
+                                                                                    index_idx, index.num_mismatch,
+                                                                                    index.idx_start, index.idx_len
+                                                                                    );
+                                        }
+                                    else
+                                        {
+                                            index_match = _find_with_shifted_mismatch( index_map, r2_seqs[ read_index ],
+                                                                                    index_idx, index.num_mismatch,
+                                                                                    index.idx_start, index.idx_len
+                                                                                    );
+                                        }
                                 }
+                            
+                            idx_match_list.push_back( index_match );
                         }
 
-                    // get the forward and the reverse indexes from the sequence, grab the id
-                    // for this index. This gives the index of the location at the sequence to increment
-                    idx1_match = _find_with_shifted_mismatch( index_map, reads[ read_index ],
-                                                               index_idx, std::get<2>( d_opts->index1_data ),
-                                                               forward_start, forward_length
-                                                             );
                     if( match_found()
                         && quality_match()
                       )
@@ -294,30 +303,27 @@ void module_demux::run( options *opts )
                                                                             seq_start,
                                                                             seq_length
                                                                           );
-
-                                    if( reverse_length == 0
+                                    if( flexible_idx_data.size() == 1
                                         && seq_match != reference_counts.end()
                                         )
                                         {
                                             std::cout << "ref match\n";
-                                            sample_id = idx1_match->second.id;
+                                            sample_id = idx_match_list[0]->second.id;
                                             seq_match->second->at( sample_id ) += 1;
                                             var_reg_match_total++;
                                             ++processed_success;
-                                            if( !d_opts->diagnostic_fname.empty() )
-                                                {
-                                                    diagnostic_map.find(
-                                                        { idx1_match->first.name, idx2_match->first.name }
-                                                    )->second.second[2] += 1;
-
-                                                }
                                         }
-                                    else if( reverse_length != 0
-                                             && seq_match != reference_counts.end()
-                                             )
+                                    else if( idx_match_list.size() > 1
+                                        && seq_match != reference_counts.end() )
                                         {
-                                            std::string concat_idx = idx1_match->first.seq
-                                                + idx2_match->first.seq;
+                                            std::string concat_idx = "";
+                                            for( const auto& index : idx_match_list )
+                                                {
+                                                    if( index != index_map.end() )
+                                                        {
+                                                            concat_idx.append( index->first.seq );
+                                                        }
+                                                }
 
                                             auto d_id = index_map.find( sequence( "", concat_idx ) );
 
@@ -327,12 +333,6 @@ void module_demux::run( options *opts )
                                                     seq_match->second->at( sample_id ) += 1;
                                                     var_reg_match_total++;
                                                     ++processed_success;
-                                                    if( !d_opts->diagnostic_fname.empty() )
-                                                        {
-                                                            diagnostic_map.find(
-                                                                { idx1_match->first.name, idx2_match->first.name }
-                                                            )->second.second[2] += 1;
-                                                        }
                                                 }
                                         }
                                     else if( seq_match == reference_counts.end()
@@ -347,7 +347,7 @@ void module_demux::run( options *opts )
                                     et_seq_search<seq_map,false> library_searcher( lib_idx, reference_counts, num_samples );
 
 
-                                    if( reverse_length == 0 )
+                                    if( flexible_idx_data.size() == 1 )
                                         {
                                             auto seq_match = library_searcher.find( reads[ read_index ],
                                                                                     std::get<2>( d_opts->seq_data ),
@@ -357,15 +357,22 @@ void module_demux::run( options *opts )
 
                                             if( seq_match != reference_counts.end() )
                                                 {
-                                                    sample_id = idx1_match->second.id;
+                                                    sample_id = idx_match_list[0]->second.id;
                                                     seq_match->second->at( sample_id ) += 1;
                                                     ++processed_success;
                                                 }
                                         }
-                                    else if( reverse_length != 0 )
+                                    else if( flexible_idx_data.size() > 1 )
                                         {
-                                            std::string concat_idx = idx1_match->first.seq
-                                                + idx2_match->first.seq;
+                                            std::string concat_idx = "";
+                                            for( const auto& index : idx_match_list )
+                                                {
+                                                    sequential_map<sequence, sample> find_end;
+                                                    if( index != find_end.end() )
+                                                        {
+                                                            concat_idx.append( index->first.seq );
+                                                        }
+                                                }
 
                                             auto d_id = index_map.find( sequence( "", concat_idx ) );
 
@@ -399,16 +406,17 @@ void module_demux::run( options *opts )
                         }
                     // record the number of records that are processed
                     ++processed_total;
+                    idx_match_list.clear();
                 }
 
             reads.clear();
             r2_seqs.clear();
-
+            
         }
-    if( !d_opts->diagnostic_fname.empty() )
-        {
-            write_diagnostic_output( d_opts, diagnostic_map );
-        }
+    // if( !d_opts->diagnostic_fname.empty() )
+    //     {
+    //         write_diagnostic_output( d_opts, diagnostic_map );
+    //     }
 
     total_time.stop();
 
@@ -416,14 +424,15 @@ void module_demux::run( options *opts )
     std::cout << processed_success << " records were found to be a match out of "
               << processed_total << " (" << ( (long double) processed_success / (long double) processed_total ) * 100
               << "%) successful.\n";
-    std::cout << std::fixed << std::setprecision( 2 )
-              << ( ( long double ) idx1_match_total / ( long double ) processed_total ) * 100.00 << "% " <<
-                        idx1_match_total << " of total records matched index 1.\n"
-              << ( ( long double ) idx2_match_total / ( long double ) processed_total ) * 100.00 << "% " <<
-                        idx2_match_total << " of total records matched index 1 + index 2.\n"
-              << ( ( long double ) var_reg_match_total / ( long double ) processed_total ) * 100.00 << "% " <<
-                        var_reg_match_total << " of total records matched index 1 + index 2 + DNA tag.\n";
-
+    // loop through index match totals
+    for( std::size_t fif_index = 0; fif_index < flexible_idx_data.size(); ++fif_index )
+        {
+            std::cout << std::fixed << std::setprecision( 2 )
+                    << ( ( long double ) index_match_totals[fif_index].second / ( long double ) processed_total ) * 100.00 << "% " <<
+                                index_match_totals[fif_index].second << " of total records matched index '" << flexible_idx_data[fif_index].idx_name << "'.\n";
+        }
+    std::cout << std::fixed << std::setprecision( 2 ) << ( ( long double ) var_reg_match_total / ( long double ) processed_total ) * 100.00 << "% " <<
+                                var_reg_match_total << " of total records matched provided indexes + DNA tag.\n";
     if( d_opts->concatemer.length() > 0 )
         {
             std::cout << "The concatemer sequence was found " << concatemer_found << " times (" <<
@@ -489,9 +498,6 @@ void module_demux::aggregate_counts( parallel_map<sequence, std::vector<std::siz
             std::string cpy = strs[ 0 ];
             current = sequence( cpy, iter->first.seq );
 
-            // we are not going to get correct results
-            // if names are not in the expected
-            // format.
             if( strs.size() != NUM_DELIMITED_ITEMS )
                 {
 
@@ -545,14 +551,20 @@ void module_demux::add_seqs_to_map( parallel_map<sequence, std::vector<std::size
         }
 }
 
-void module_demux::write_diagnostic_output( options_demux* d_opts, std::map<std::pair<std::string,std::string>,
+void module_demux::write_diagnostic_output( options_demux* d_opts, std::map<std::vector<std::string>,
              std::pair<std::string,std::vector<std::size_t>>>& diagnostic_map )
 {
     std::ofstream outfile( d_opts->diagnostic_fname, std::ios::out );
-    std::map<std::pair<std::string,std::string>,
+    std::map<std::vector<std::string>,
              std::pair<std::string,std::vector<std::size_t>>>::iterator sample_iter = diagnostic_map.begin();
-    std::string header = "Sample name\tIndex1 Matches\tIndex Pair Matches";
-    // Include Variable Region Matches column if ref dependent.
+    std::string header = "Sample name\t";
+    std::vector<std::string> header_index_cols;
+    // this should be for the size of the index columns? I'm looking to reserve this vector to the size of the num of indexes.
+    header_index_cols.reserve( diagnostic_map.begin()->second.second.size() );
+    // for the length of indexes provided by fif list
+    
+        // add to header the first index name to every column index from end to beginning.
+        // the total range should be reduced by 1. So the first column in this list doesn't have the second index concatenated to it.
     if( !d_opts->library_fname.empty() )
         {
             header.append( "\tDNA Tag Matches" );
@@ -640,40 +652,48 @@ void module_demux::_zero_vector( std::vector<std::size_t>* vec )
 
 void module_demux::create_index_map( sequential_map<sequence, sample>& map,
                                      std::vector<sequence>& index_seqs,
-                                     std::vector<sample>& samplelist
+                                     std::vector<sample>& samplelist,
+                                     std::vector<flex_idx>& flexible_idx_data
                                    )
 {
-    sequential_map<std::string, std::string> st_table;
-    std::string concat_seq = "";
-    std::string seq1       = "";
-    std::string seq2       = "";
-
-    unsigned int index = 0;
-
-    // Populate map with first sample as default filler data.
-    for( index = 0; index < index_seqs.size(); ++index )
+    // sequential_map<std::string, std::string> st_table;
+    std::unordered_map<std::string, std::string> updated_index_seqs;    
+    for( auto& index : index_seqs )
         {
-            st_table[ index_seqs[ index ].name ] = index_seqs[ index ].seq;
-            for( const auto& sample : samplelist )
+            bool index_seq_found = false;
+            std::size_t curr_sample = 0;
+            // while the current index seq id has not been found in a sample and while there are still samples
+            while( !index_seq_found && curr_sample < samplelist.size() )
                 {
-                    if( sample.get_first_id().compare(index_seqs[index].name) == 0
-                        || sample.get_second_id().compare(index_seqs[index].name) == 0 )
+                    // look at each string id and compare
+                    for( auto& id : samplelist[curr_sample].string_ids )
                         {
-                             map[ sequence( index_seqs[ index ].name,
-                                            index_seqs[ index ].seq)
-                                ] = samplelist[ 0 ];
+                            // if the current id does match the index seqs id
+                            if( index.name.compare( id ) == 0 )
+                                {
+                                    // mark as the index seq was found
+                                    index_seq_found = true;
+                                    // add the index seq to the updated list (order doesnt matter)
+                                    updated_index_seqs.emplace( index.name, index.seq );
+                                }
                         }
+                    curr_sample++;
                 }
         }
-
-    for( index = 0; index < samplelist.size(); ++index )
+    // now we need to add partial and complete match refs.
+        
+    // future function: add_match_references -> add the partial matches and complete matches by creating 
+    for( std::size_t curr_sample = 0; curr_sample < samplelist.size(); ++curr_sample )
         {
-            // This is now N sequences.
-            seq1 = st_table[ samplelist[ index ].get_first_id() ];
-            seq2 = st_table[ samplelist[ index ].get_second_id() ];
-            concat_seq = seq1 + seq2;
-
-            map[ sequence( "", concat_seq ) ] = samplelist[ index ];
+            std::string concat_seq = "";
+            // for the current index in the fid
+            for( std::size_t curr_idx = 0; curr_idx < samplelist[curr_sample].string_ids.size(); ++curr_idx )
+                {
+                    // obtain the barcode for the current index for the current sample and add to sequence concat to create element.
+                    auto sample_seq = updated_index_seqs.find( samplelist[curr_sample].string_ids[curr_idx] )->second;
+                    concat_seq.append( sample_seq );
+                    map[ sequence( "", concat_seq ) ] = samplelist[ curr_sample ];
+                }
         }
 }